Separate ModelParamTraversal into traverse_util.
flax.optim is on a long-term deprecation plan, move the more
generally useful ModelParamTraversal out of optim and into
traverse_util where it naturally belongs.
